backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-socketio==5.5.1
PyJWT==2.10.1
werkzeug==3.1.3
mutagen==1.47.0
